latexml_post

Module collector

Abstract collector base for content generation processors.

Port of LaTeXML::Post::Collector (113 lines of Perl). Base class for processors that collect information from multiple documents and build derived content (indexes, bibliographies, etc.). Supports splitting collected content into sub-documents by initial letter.

Structs

Collector

Abstract collector post-processor.

Functions

get_page_name

Compute a page name for a sub-collection document.

make_sub_collection_documents

Given collected content broken into portions by initial letter, fill in the main document with the first sub-collection, and create new documents for the rest.